refactor: reorganize supabase client helpers
